feat: handle finish
