+++ src/ws_server/modules/reducer/game/index.ts
@@ -1,0 +1,1 @@
+export * from './reducer.js';
